% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_month_to_date.R
\name{fiscal_month_to_date}
\alias{fiscal_month_to_date}
\title{Convert fiscal month to year-month-day date format}
\usage{
fiscal_month_to_date(fiscal_month, day = NULL)
}
\arguments{
\item{fiscal_month}{(character) Date in fiscal month format (e.g., "FY22 - Apr").}

\item{day}{(character) Which day to use in date? If NULL, defaults to last day of the month.}
}
\value{
(date) lubridate date object in "yyyy-mm-dd" format.
}
\description{
Convert fiscal month to year-month-day date format
}
