% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_mgmt_lvl_hierarchy.R
\name{fix_mgmt_lvl_hierarchy}
\alias{fix_mgmt_lvl_hierarchy}
\title{Fix management level hierarchy (L2/L3/L4)}
\usage{
fix_mgmt_lvl_hierarchy(dat)
}
\arguments{
\item{dat}{Data containing management level variables, which should be called
"management_chain_level_2", "management_chain_level_3", and ""management_chain_level_4".
Currently doesn't support management levels beyond that.}
}
\value{
Original data from with management chain variables fixed.
}
\description{
Fix management level hierarchy (L2/L3/L4)
}
