% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_us_minority.R
\name{flag_us_minority}
\alias{flag_us_minority}
\title{Flag U.S. Minority}
\usage{
flag_us_minority(race_ethnicity)
}
\arguments{
\item{race_ethnicity}{Character vector containing the employee race/ethnicity data.}
}
\description{
Flag U.S. Minority
}
