% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmi_color_palettes.R
\name{gmi_cols}
\alias{gmi_cols}
\title{Function to extract gmi colors as hex codes}
\usage{
gmi_cols(...)
}
\arguments{
\item{...}{Character names of gmi_colors}
}
\description{
Function to extract gmi colors as hex codes
}
