% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRAnalyticsTools-package.R
\docType{package}
\name{HRAnalyticsTools-package}
\alias{HRAnalyticsTools}
\alias{HRAnalyticsTools-package}
\title{HRAnalyticsTools: Tools for the GMI HR Analytics Team}
\description{
Functions and tools for the GMI HR Analytics team.
}
\keyword{internal}
