% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_management_level.R
\name{recode_management_level}
\alias{recode_management_level}
\title{Recode management level, optionally collapsing officer groups into a single category}
\usage{
recode_management_level(management_level, collapse_officer = TRUE)
}
\arguments{
\item{management_level}{Character vector of management levels.}

\item{collapse_officer}{(Logical) Collapse VP, Senior VP, SLT, and CEO into a single group? Default is TRUE.}
}
\description{
Recode management level, optionally collapsing officer groups into a single category
}
