% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_performance_rating.R
\name{recode_performance_rating}
\alias{recode_performance_rating}
\title{Recode performance rating}
\usage{
recode_performance_rating(performance_rating)
}
\arguments{
\item{performance_rating}{A character vector of performance ratings in all historical formats.}
}
\value{
A numeric vector of performance ratings on the 1-3 scale.
}
\description{
Recode performance rating
}
