% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_race_ethnicity.R
\name{recode_race_ethnicity}
\alias{recode_race_ethnicity}
\title{Recode race/ethnicity variable}
\usage{
recode_race_ethnicity(race_ethnicity)
}
\arguments{
\item{race_ethnicity}{(Character) Vector containing the raw race/ethnicity field from Workday.}
}
\description{
Recode race/ethnicity variable
}
