% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmi_color_palettes.R
\name{scale_fill_gmi}
\alias{scale_fill_gmi}
\title{Fill scale constructor for gmi colors}
\usage{
scale_fill_gmi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gmi_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for gmi colors
}
\examples{
ggplot2::ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_boxplot() +
  scale_fill_gmi()
}
