% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gmi.R
\name{theme_gmi}
\alias{theme_gmi}
\title{Basic GMI theme}
\usage{
theme_gmi(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base text size.}

\item{base_family}{Base text family.}

\item{base_line_size}{Base line size.}

\item{base_rect_size}{Base rectangle size.}
}
\description{
Basic GMI theme
}
